Migrate to Bazel build system and update CI/validation workflows (#785)
* Migrate to Bazel build system and update CI/validation workflows

* fix(workspace): remove incorrect plan check during workspace publish

* ci: fix setup-bazel action versions and validate disk-cache key

* docs: remove nix support and migrate fmt/clippy jobs to cargo

* ci: pin rust toolchain version to 1.91.1 across jobs

* tests: resolve decapod binary using resolve_decapod_bin helper to support Bazel runfiles execution

* build: bump Rust toolchain to 1.96.1 and update test binary paths in BUILD.bazel

* build: fix clippy warnings and collapsible_if in tests under Rust 1.96.1

* ci: update rust-cache shared-key to force clean clippy run

* tests: fix remaining un-canonicalized decapod path in agent_rpc_suite.rs

* build: add .decapod config and contract metadata files to integration test data dependencies

---------

+++ src/core/todo.rs
@@ -1276,7 +1276,7 @@ pub fn cleanup_stale_agent_assignments(
                     )
                     .map_err(error::DecapodError::RusqliteError)?;
                     sync_legacy_owner_column(conn, &task_id)?;
-                    released_count += changed as usize;
+                    released_count += changed;
 
                     let ev = TodoEvent {
                         ts: ts.clone(),
